requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
pytest
Flask-Marshmallow